+++ src/main/java/de/raidcraft/skills/Scoreboards.java
@@ -81,7 +81,7 @@ public final class Scoreboards {
                 if (objective.getName().startsWith(BukkitUserInterface.HEALTH_OBJECTIVE)) {
                     for (Player player : Bukkit.getOnlinePlayers()) {
                         if (player.getHealth() > 0) {
-                            objective.getScore(player).setScore(player.getHealth());
+                            objective.getScore(player).setScore((int) player.getHealth());
                         }
                     }
                     break;
